Time: 2 ms (100.00%) | Memory: 72.7 MB (58.37%) - LeetSync

+++ 1803-average-waiting-time/average-waiting-time.java
@@ -6,6 +6,6 @@ class Solution {
             time = (time < arr[0]) ? arr[0] + arr[1] : time + arr[1];
             waiting += time - arr[0];    
         }
-        return (double) waiting / customers.length;
+        return waiting / customers.length;
     }
 }
